+++ R/api.R
@@ -16,8 +16,8 @@ api_build <- function(queue) {
   api$handle(endpoint_plotting_metadata())
   api$handle(endpoint_download_spectrum(queue))
   api$handle(endpoint_download_spectrum_head(queue))
-  api$handle(endpoint_download_summary(queue))
-  api$handle(endpoint_download_summary_head(queue))
+  api$handle(endpoint_download_coarse_output(queue))
+  api$handle(endpoint_download_coarse_output_head(queue))
   api$handle(endpoint_hintr_version())
   api$handle(endpoint_hintr_worker_status(queue))
   api$handle(endpoint_hintr_stop(queue))
@@ -317,17 +317,17 @@ endpoint_download_spectrum_head <- function(queue) {
                               validate = FALSE)
 }
 
-endpoint_download_summary <- function(queue) {
+endpoint_download_coarse_output <- function(queue) {
   pkgapi::pkgapi_endpoint$new("GET",
-                              "/download/summary/<id>",
-                              download_summary(queue),
+                              "/download/coarse-output/<id>",
+                              download_coarse_output(queue),
                               returning = pkgapi::pkgapi_returning_binary())
 }
 
-endpoint_download_summary_head <- function(queue) {
+endpoint_download_coarse_output_head <- function(queue) {
   pkgapi::pkgapi_endpoint$new("HEAD",
-                              "/download/summary/<id>",
-                              download_summary(queue),
+                              "/download/coarse-output/<id>",
+                              download_coarse_output(queue),
                               returning = returning_binary_head(),
                               validate = FALSE)
 }